Remove allowing Testlib to be imported in non-test programs
